+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## 2.0.0 - 2017-06-03
 
+- Add DOcker instructions
 - Add Benchmark reults (1K requests => 2.5s)
 - Restart server if file changes
 - Refactored library to use EventEmitters